Omit lots of definitions from mips-unknown-linux-uclibc
